+++ crates/sncast/src/lib.rs
@@ -2,9 +2,7 @@ use std::num::{NonZeroU8, NonZeroU16};
 
 use crate::helpers::account::{check_account_exists, get_account_from_devnet, is_devnet_account};
 use crate::helpers::configuration::CastConfig;
-use crate::helpers::constants::{
-    DEFAULT_ACCOUNTS_FILE, DEFAULT_STATE_FILE_SUFFIX, WAIT_RETRY_INTERVAL, WAIT_TIMEOUT,
-};
+use crate::helpers::constants::{DEFAULT_ACCOUNTS_FILE, WAIT_RETRY_INTERVAL, WAIT_TIMEOUT};
 use crate::helpers::rpc::RpcArgs;
 use crate::response::errors::SNCastProviderError;
 use anyhow::{Context, Error, Result, anyhow, bail, ensure};
@@ -51,7 +49,6 @@ use url::Url;
 
 pub mod helpers;
 pub mod response;
-pub mod state;
 
 use crate::helpers::ledger;
 use crate::response::ui::UI;
@@ -939,11 +936,6 @@ macro_rules! apply_optional_fields {
             value
         }
     };
-}
-
-#[must_use]
-pub fn get_default_state_file_name(script_name: &str, chain_id: &str) -> String {
-    format!("{script_name}_{chain_id}_{DEFAULT_STATE_FILE_SUFFIX}")
 }
 
 #[cfg(test)]